Replace cheat implementation with own implementation.
Breaks the pattern up in sub-patterns.
A subpattern has a mask that can either be a fixed character or a wildcard matching any character.
Has a flag indicating whether the subpatern should match exactly 1 character, or any number of characters (0 or more).
The pattern always has to match the entire input string.