Make pattern loop leading.
We need to match all the patterns. The previous implementation had to have an ugly loop at the end to skip over variable length pattern. That's ugly. Also semantically, "a character matches 0 or more patterns" was not true. A pattern matches 0 or more characters.
This change makes the pattern loop the outer loop, and matches input characters in the inner loop. This feels more natural and gets rid of the ugly end loop.