Do some pre-fixes for backtracking
Iteration should start at 0. That's the pattern matching itself. How many it should match is up to the pattern properties, limited by its min and max properties.
A match has a length, not an interation per se. It's up to the subpattern to decide if it can do another iteration.
Currently that is from min to max. Later on, max should be lowered on each next backtracking attempt.
When implementing non-greedy quantifiers one day, the matcher should increase the min instead of decreasing the max, when backtracking.